requirements.txt
fire==0.5.0
langchain==0.0.144
PyMuPDF==1.22.1
python-dotenv==1.0.0
openai==0.27.4
chromadb==0.3.21
tiktoken==0.3.3